modules: Fix issue with unloading modules bound to OnMessage
When unloaded, modules attached to OnMessage events would still be part
of the list used by FOREACH_MOD to send events to, leading to a
segfault.

This changes the behavior of load and unload to place new modules at the
*start* of the event handler arrays, leading them to already be
processed by the time the bot gets to the ModuleManager.